requirements.txt
google.cloud.logging
pandas
jira>=3.5.1
google-cloud-bigquery
pyarrow